+++ CHANGELOG.md
@@ -36,6 +36,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+
+- (fix) xap-153 | packages/instant-dapps 1.0.4 packages/widget 1.0.7 package/i18n 1.0.5 | Wormhole temporary banner
 - (fix) xap-150 | apps/mission 1.0.34 package/instant-dapps 1.0.3 package/ widgets 1.0.6 pages 1.1.7 | add TFM instant dapp
 
 ## 2.0.10 - 2024-03-11